requirements.txt
joblib==1.2.0
matplotlib==3.7.1
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
requests==2.29.0
scikit-learn==1.2.2
scipy==1.10.1
streamlit==1.22.0
streamlit-lottie==0.0.3
toml==0.10.2
xlrd==2.0.1
